% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance00.R
\docType{data}
\name{distance00}
\alias{distance00}
\title{This is data that contains distance between capitals for 2000-2012 years years
(Kristian Skrede Gleditsch)}
\format{An object of class \code{data.frame} with 187 rows and 187 columns.}
\usage{
distance00
}
\description{
This is data that contains distance between capitals for 2000-2012 years years
(Kristian Skrede Gleditsch)
}
\author{
Zhanna Terechshenko \email{zhanna.terechshenko@gmail.com}
}
\keyword{datasets}
