% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance75.R
\docType{data}
\name{distance75}
\alias{distance75}
\title{This is data that contains distance between capitals for 1975-1989 years
(Kristian Skrede Gleditsch)}
\format{An object of class \code{data.frame} with 149 rows and 149 columns.}
\usage{
distance75
}
\description{
This is data that contains distance between capitals for 1975-1989 years
(Kristian Skrede Gleditsch)
}
\author{
Zhanna Terechshenko \email{zhanna.terechshenko@gmail.com}
}
\keyword{datasets}
