% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\docType{data}
\name{edgelist}
\alias{edgelist}
\title{This is data that contains edgelist}
\format{An object of class \code{data.frame} with 48304 rows and 4 columns.}
\usage{
edgelist
}
\description{
This is data that contains edgelist
}
\author{
Zhanna Terechshenko \email{zhanna.terechshenko@gmail.com}
}
\keyword{datasets}
