% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVs00.R
\docType{data}
\name{IVs00}
\alias{IVs00}
\title{This is data that contains independent variables for 2000-2012 years}
\format{An object of class \code{data.frame} with 2431 rows and 15 columns.}
\usage{
IVs00
}
\description{
This is data that contains independent variables for 2000-2012 years
}
\author{
Zhanna Terechshenko \email{zhanna.terechshenko@gmail.com}
}
\keyword{datasets}
