% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVs75.R
\docType{data}
\name{IVs75}
\alias{IVs75}
\title{This is data that contains independent variables for 1975-1989 years}
\format{An object of class \code{data.frame} with 2235 rows and 14 columns.}
\usage{
IVs75
}
\description{
This is data that contains independent variables for 1975-1989 years
}
\author{
Zhanna Terechshenko \email{zhanna.terechshenko@gmail.com}
}
\keyword{datasets}
