% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinf_dynamic.R
\name{netinf_dynamic}
\alias{netinf_dynamic}
\title{A Function to create estimates for network inference for each year (25 year window)}
\usage{
netinf_dynamic(data, edges = NULL, window = NULL, startYear = NULL,
  endYear = NULL)
}
\arguments{
\item{data.frame}{Final dataset used in the project}
}
\value{
data.frame with estimates
}
\description{
A Function to create estimates for network inference for each year (25 year window)
}
