% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm_2000.R
\name{tergm_2000}
\alias{tergm_2000}
\title{A Function to run TERGM for the post-Cold War years}
\usage{
tergm_2000(ivs00, edgelist, distance00, tradelist00)
}
\arguments{
\item{data.frame}{with independent variables for 2000-2012 years,
data.frame with netinf estimates, two dataframes with distance and trade estimates}
}
\value{
list
}
\description{
A Function to run TERGM for the post-Cold War years
}
