% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradelist00.R
\docType{data}
\name{tradelist00}
\alias{tradelist00}
\title{This is data on trade flows between states (as defined by the Correlates of War project) for the period 2000-2012 years}
\format{An object of class \code{list} of length 13.}
\usage{
tradelist00
}
\description{
This is data on trade flows between states (as defined by the Correlates of War project) for the period 2000-2012 years
}
\author{
Zhanna Terechshenko \email{zhanna.terechshenko@gmail.com}
}
\keyword{datasets}
