% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradelist75.R
\docType{data}
\name{tradelist75}
\alias{tradelist75}
\title{This is data on trade flows between states (as defined by the Correlates of War project) for the period 1975-1989 years}
\format{An object of class \code{list} of length 15.}
\usage{
tradelist75
}
\description{
This is data on trade flows between states (as defined by the Correlates of War project) for the period 1975-1989 years
}
\author{
Zhanna Terechshenko \email{zhanna.terechshenko@gmail.com}
}
\keyword{datasets}
