% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tergm_2000.R
\name{tergm_2000}
\alias{tergm_2000}
\title{A Function to run TERGM for for 2000-2012 years}
\usage{
tergm_2000(ivs00, edgelist)
}
\arguments{
\item{data.frame}{Final dataset used in the project}
}
\value{
data.frame with independent variables for 2000-2012 years, data.frame with netinf estimates
}
\description{
A Function to run TERGM for the post-Cold War years
}
